refactor: split the cluster properties sync onto two controllers
- one that sync off the domain base prefix from CS
- the existing one is now repurposed to read off HostedCluster cr and sync the remaining properties. It does so when the values changed i.e existing != observed
