+++ backend/pkg/controllers/clusterpropertiescontroller/cluster_base_domain_prefix_sync.go
@@ -1,0 +1,113 @@
+// Copyright 2026 Microsoft Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package clusterpropertiescontroller
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"k8s.io/apimachinery/pkg/api/equality"
+
+	"github.com/Azure/ARO-HCP/backend/pkg/controllers/controllerutils"
+	"github.com/Azure/ARO-HCP/backend/pkg/informers"
+	"github.com/Azure/ARO-HCP/backend/pkg/listers"
+	controllerutil "github.com/Azure/ARO-HCP/internal/controllerutils"
+	"github.com/Azure/ARO-HCP/internal/database"
+	unionkubeapplierinformers "github.com/Azure/ARO-HCP/internal/database/unioninformers/kubeapplier"
+	"github.com/Azure/ARO-HCP/internal/ocm"
+	"github.com/Azure/ARO-HCP/internal/utils"
+)
+
+// clusterBaseDomainPrefixSyncer synchronizes CustomerProperties.DNS.BaseDomainPrefix from
+// Cluster Service to Cosmos DB when the field is unset.
+type clusterBaseDomainPrefixSyncer struct {
+	cooldownChecker      controllerutil.CooldownChecker
+	resourcesDBClient    database.ResourcesDBClient
+	clusterServiceClient ocm.ClusterServiceClientSpec
+}
+
+var _ controllerutils.ClusterSyncer = (*clusterBaseDomainPrefixSyncer)(nil)
+
+// NewClusterBaseDomainPrefixSyncController creates a controller that synchronizes
+// CustomerProperties.DNS.BaseDomainPrefix from Cluster Service to Cosmos DB.
+func NewClusterBaseDomainPrefixSyncController(
+	resourcesDBClient database.ResourcesDBClient,
+	clusterServiceClient ocm.ClusterServiceClientSpec,
+	activeOperationLister listers.ActiveOperationLister,
+	informers informers.BackendInformers,
+	kubeApplierInformers *unionkubeapplierinformers.UnionKubeApplierInformers,
+) controllerutils.Controller {
+	syncer := &clusterBaseDomainPrefixSyncer{
+		cooldownChecker:      controllerutils.DefaultActiveOperationPrioritizingCooldown(activeOperationLister),
+		resourcesDBClient:    resourcesDBClient,
+		clusterServiceClient: clusterServiceClient,
+	}
+
+	return controllerutils.NewClusterWatchingController(
+		"ClusterBaseDomainPrefixSync",
+		resourcesDBClient,
+		informers,
+		kubeApplierInformers,
+		5*time.Minute,
+		syncer,
+	)
+}
+
+func (c *clusterBaseDomainPrefixSyncer) CooldownChecker() controllerutil.CooldownChecker {
+	return c.cooldownChecker
+}
+
+func (c *clusterBaseDomainPrefixSyncer) SyncOnce(ctx context.Context, key controllerutils.HCPClusterKey) error {
+	logger := utils.LoggerFromContext(ctx)
+
+	clusterCRUD := c.resourcesDBClient.HCPClusters(key.SubscriptionID, key.ResourceGroupName)
+	existingCluster, err := clusterCRUD.Get(ctx, key.HCPClusterName)
+	if database.IsNotFoundError(err) {
+		return nil
+	}
+	if err != nil {
+		return utils.TrackError(fmt.Errorf("failed to get Cluster: %w", err))
+	}
+
+	if len(existingCluster.CustomerProperties.DNS.BaseDomainPrefix) > 0 {
+		return nil
+	}
+
+	if existingCluster.ServiceProviderProperties.ClusterServiceID == nil ||
+		len(existingCluster.ServiceProviderProperties.ClusterServiceID.String()) == 0 {
+		return nil
+	}
+
+	originalCluster := existingCluster.DeepCopy()
+
+	csCluster, err := c.clusterServiceClient.GetCluster(ctx, *existingCluster.ServiceProviderProperties.ClusterServiceID)
+	if err != nil {
+		return utils.TrackError(fmt.Errorf("failed to get cluster from Cluster Service: %w", err))
+	}
+
+	existingCluster.CustomerProperties.DNS.BaseDomainPrefix = csCluster.DomainPrefix()
+
+	if equality.Semantic.DeepEqual(originalCluster, existingCluster) {
+		return nil
+	}
+
+	if _, err := clusterCRUD.Replace(ctx, existingCluster, nil); err != nil {
+		return utils.TrackError(fmt.Errorf("failed to replace Cluster: %w", err))
+	}
+
+	logger.Info("synced cluster base domain prefix")
+	return nil
+}
